+++ src/components/LoginForms/EmailPassword/index.ts
@@ -1,0 +1,1 @@
+export * from './EmailPassword'
